Rename folder to FeatureCollection, set Feature to be IResource and move NavPlace to StructureBase
