+++ src/IIIF/IIIF/Serialisation/Deserialisation/ResourceConverter.cs
@@ -3,6 +3,7 @@ using IIIF.Auth.V2;
 using IIIF.ImageApi.V2;
 using IIIF.ImageApi.V3;
 using IIIF.Presentation.V3.Content;
+using IIIF.Presentation.V3.FeatureCollection;
 using Newtonsoft.Json.Linq;
 
 namespace IIIF.Serialisation.Deserialisation;
@@ -53,6 +54,7 @@ public class ResourceConverter : ReadOnlyConverter<IResource>
                     nameof(Sound) => new Sound(),
                     nameof(Video) => new Video(),
                     nameof(Image) => new Image(),
+                    nameof(Feature) => new Feature(),
                     _ => null
                 };
         }
